requirements.txt
cryptography>=41.0.0
PyJWT>=2.8.0
pyinstaller>=6.0.0